+++ src/VirtoCommerce.OrdersModule.Core/Model/Search/OrderAggregationLabel.cs
@@ -1,0 +1,8 @@
+﻿namespace VirtoCommerce.OrdersModule.Core.Model.Search
+{
+    public class OrderAggregationLabel
+    {
+        public string Language { get; set; }
+        public string Label { get; set; }
+    }
+}
